+++ ui/components/settings/setup.mjs
@@ -26,11 +26,17 @@ const withoutTMP = (obj) => ({ ...obj, TMP: undefined })
 /*
  * Displays configuration validation
  */
-const ShowConfigurationValidation = ({ deploy, validationReport, toggleValidate }) => (
+const ShowConfigurationValidation = ({ deploy, validationReport, toggleValidate}) => (
   <>
     {validationReport ? (
       <>
         <h3>Validation Results</h3>
+        {validationReport.status === 400 && validationReport.schema_violation ? (
+        <Popout tip>
+          <h4>{validationReport.title}</h4>
+          <p>{validationReport.schema_violation}</p>
+        </Popout>
+        ) : null }
         <SettingsReport report={validationReport} />
         {validationReport.valid ? (
           <button className="btn btn-accent btn-lg w-full mt-4" onClick={deploy}>
@@ -142,7 +148,7 @@ export const SetupWizard = ({ preload = {}, validate = false }) => {
     }
     setValidateView(!validateView)
   }
-  const formEl = cluster(mSettings, toggleValidate).children.setup.form
+  const formEl = cluster(mSettings, toggleValidate, update).children.setup.form
 
   return (
     <div className="w-full max-w-2xl mx-auto">