+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [api] Allow the export of key data
 - [core] Allow the export of key data
 - [ui] Allow the export of key data
+- [ui] Allow loading key data file in setup wizard
 - [watcher] We have enabled the watcher service which was disabled due to SASL
 
 ### Fixed